Start creating OAuth flow
A new LoginView has been added which contains the view for getting a user's
domain, which should open the OAuth page respectively. The starlight URI
scheme has also been registered.

Further commits will need to register the application and properly deep link
the OAuth to store the token correctly as per the documentation.