Expand Explore page to own view with search capabilities
The Explore page is now in its own view with the searchable modifier enabled.
I am unsure of how we can use the search bar to populate a search request,
but this is still a work in progress. Search suggestions for users and hashtags
appear as normal.

Additionally, before trends load in, a placeholder with five tags is inserted to
prevent massive UI jittering. The trending tags localization has also been fixed
by writing an explicit header with localizable text.