Do stuff with macOS + iPadOS
Some touch ups have been made to the standard navigation layout for the macOS and
iPadOS clients to match the current iOS counterparts. Trending tags appear in the side
bar along with the new Home, Network, and Messages tabs. Older sidebar actions have
been removed to better follow the new design.

Some pieces of text have been given new localizations in English.

The settings view from the Settings tab on iOS has been ported to macOS as the
Preferences window.

The login window will appear if the user is not logged in, though the OAuth flow
seems to disregard whether or not the OAuth status has changed, and it opens a
new instance on macOS.

The profile card now has three different configurable sizes for the profile image.

Entitlements for macOS have been added with network configurations.