Refactor TimelineMasterDetailView, TimelineScrollViewCompatible
Both of these now use an underlying TimelineViewable, which is responsible for fetching the data
and handling async state representation with views. TimelineViewable merely requires the
developer to pass in a ViewBuilder which specifies what the loaded result will look like, instead of
rewriting the code for representing all of the states.

This was necessary to get ProfileView to display posts from a user's profile, since the scope in
TimelineViewable has been adjusted for this. The ProfileView's TimelineScrollViewCompatible
seems to be suffering from a ForEach, issue, however:

ForEach<Range<Int>, Int, AnyView> count (157) != its initial count (1). `ForEach(_:content:)`
should only be used for *constant* data. Instead conform data to `Identifiable` or use
`ForEach(_:id:content:)` and provide an explicit `id`!

Which shouldn't be an issue because we're using the status's ID, but it's not being registered.
I suspect the nesting of this inside of the View may be causing some of the issues here, since
it's not being affected on the Home and Network pages.