Refactor blurb section of ProfileView
The blurb section of the ProfileView uses reusable methods and variables that will construct
the blurbs themselves, trimming down the size of the code and making it more readable.

It also appears that the action for the blurb button is broken; i.e., it doesn't do anything. This
should be investigated.